'refactor: change code structure
- create components folder
- place components logic in the folder components
- create index file for components as a single source
- create styles folder
- place styles files in styles folder
- import components in main index file
'
