+++ nav-bar-sticky/src/components/About.tsx
@@ -1,5 +1,5 @@
-import './About.css';
+import '../styles/About.css';
 
 const About = () => <div>Welcome to the section About us!</div>;
 
-export default About;
+export { About };
