+++ nav-bar-sticky/src/components/Offer.tsx
@@ -1,5 +1,5 @@
-import './Offer.css';
+import '../styles/Offer.css';
 
 const Offer = () => <div>Welcome to the Offer section!</div>;
 
-export default Offer;
+export { Offer };
