+++ nav-bar-sticky/src/components/index.tsx
@@ -1,0 +1,6 @@
+export { About } from './About';
+export { Contact } from './Contact';
+export { ErrorPage } from './ErrorPage';
+export { Navigation } from './Navigation';
+export { Offer } from './Offer';
+export { RootComponent } from './RootComponent';
